Create CommonChild.cs
A string is said to be a child of a another string if it can be formed by deleting 0 or more characters from the other string. Letters cannot be rearranged. Given two strings of equal length, what's the longest string that can be constructed such that it is a child of both?

Example
s1="ABCD"
s2="ABDC"

These strings have two children with maximum length 3, ABC and ABD. They can be formed by eliminating either the D or C from both strings. Return: 3. 

https://www.hackerrank.com/challenges/common-child/problem?isFullScreen=true&h_r=next-challenge&h_v=zen